pyproject.toml
[tool.poetry]
name = "sharemarket-analysis"
version = "1.0.0"
description = "Indian Stock Market Analysis Tool for Intraday and Swing Trading"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.5.0"
numpy = "^1.21.0"
yfinance = "^0.2.0"
requests = "^2.25.0"
python-dotenv = "^0.19.0"
loguru = "^0.6.0"
scikit-learn = "^1.0.0"
matplotlib = "^3.5.0"
seaborn = "^0.11.0"
plotly = "^5.0.0"
ta = "^0.10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^22.0.0"
flake8 = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
sharemarket = "main:main"